Creating the #can_move? method to check for valid movements
This method allows to check if the user is not placing the desired piece
on top of another piece of the same color, and allowing only to place
them either on empty squares or on adversary pieces. Still need to
implement check for valid moves like diagonal, horizontal and other placement.