Adding the correct selection functionality
Making sure the piece/square is only selected when it has a piece. This
prevents the user from selecting an empty square.
Also adding the functionality of the piece being moved to the new square

Still have to implement the allowing of selecting a piece according to
whose turn it is and maybe refactor.